Fix vagrant installation for archlinux

+++ recipes/virtualization.sh
@@ -17,19 +17,21 @@ __get_vagrant_latest_version() {
 }
 
 kickstart.package.install_vagrant() {
-    kickstart.command_exists vagrant || {
-        package=$(__get_vagrant_latest_version)
-        version=$(sed 's/[a-z_]*//g' <<< "$package")
+    if ! kickstart.command_exists vagrant; then
+        if kickstart.os.is "Ubuntu"; then
+            package=$(__get_vagrant_latest_version)
+            version=$(sed 's/[a-z_]*//g' <<< "$package")
 
-        wget -q "https://releases.hashicorp.com/vagrant/$version/${package}_x86_64.deb"
-        wget -q "https://releases.hashicorp.com/vagrant/$version/${package}_SHA256SUMS"
-        wget -q "https://releases.hashicorp.com/vagrant/$version/${package}_SHA256SUMS.sig"
+            wget -q "https://releases.hashicorp.com/vagrant/$version/${package}_x86_64.deb"
+            wget -q "https://releases.hashicorp.com/vagrant/$version/${package}_SHA256SUMS"
+            wget -q "https://releases.hashicorp.com/vagrant/$version/${package}_SHA256SUMS.sig"
 
-        kickstart.gpg_verify "${package}_SHA256SUMS.sig" "${package}_SHA256SUMS"
-        sha256sum -c <(grep "64.deb" "${package}_SHA256SUMS")
+            kickstart.gpg_verify "${package}_SHA256SUMS.sig" "${package}_SHA256SUMS"
+            sha256sum -c <(grep "64.deb" "${package}_SHA256SUMS")
 
-        dpkg -i "./${package}_x86_64.deb"
+            dpkg -i "./${package}_x86_64.deb"
 
-        rm -rf vagrant*
-    }
+            rm -rf vagrant*
+        fi
+    fi
 }
